chore: 패키지 이동
- MainXX 클래스는 ui/main 패키지로 묶음
- 기본 화면의 네비게이션 Tab들은 ui/tabs 패키지로 묶음
