+++ app/src/main/java/com/woowa/banchan/ui/main/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.woowa.banchan
+package com.woowa.banchan.ui.main
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.woowa.banchan.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
