+++ app/src/main/java/com/woowa/banchan/ui/tabs/home/HomeFragment.kt
@@ -1,19 +1,24 @@
-package com.woowa.banchan.ui.maindish
+package com.woowa.banchan.ui.tabs.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.woowa.banchan.databinding.FragmentHomeBinding
 
-class MainDishFragment : Fragment() {
+class HomeFragment : Fragment() {
+
+    private var _binding: FragmentHomeBinding? = null
+    private val binding: FragmentHomeBinding get() = requireNotNull(_binding)
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return super.onCreateView(inflater, container, savedInstanceState)
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
