+++ app/src/main/java/com/woowa/banchan/ui/tabs/maindish/MainDishFragment.kt
@@ -1,24 +1,19 @@
-package com.woowa.banchan.ui.home
+package com.woowa.banchan.ui.tabs.maindish
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.woowa.banchan.databinding.FragmentHomeBinding
 
-class HomeFragment : Fragment() {
-
-    private var _binding: FragmentHomeBinding? = null
-    private val binding: FragmentHomeBinding get() = requireNotNull(_binding)
+class MainDishFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
-        return binding.root
+        return super.onCreateView(inflater, container, savedInstanceState)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
