+++ app/src/main/java/com/woowa/banchan/ui/tabs/side/SideFragment.kt
@@ -1,4 +1,4 @@
-package com.woowa.banchan.ui.soup
+package com.woowa.banchan.ui.tabs.side
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 
-class SoupFragment : Fragment() {
+class SideFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
