+++ app/src/main/java/com/woowa/banchan/ui/tabs/soup/SoupFragment.kt
@@ -1,4 +1,4 @@
-package com.woowa.banchan.ui.side
+package com.woowa.banchan.ui.tabs.soup
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 
-class SideFragment : Fragment() {
+class SoupFragment : Fragment() {
 
     override fun onCreateView(
         inflater: LayoutInflater,
